# backend/internal/server/router.go
package server

import (
	"context"
	"fmt"
	"net/http"
	"strings"

	"ops-system/backend/internal/config"
	"ops-system/backend/internal/grafana"
	"ops-system/backend/internal/handler"
	"ops-system/backend/internal/helm"
	"ops-system/backend/internal/idempotency"
	integrationpkg "ops-system/backend/internal/integration"
	"ops-system/backend/internal/k8s"
	"ops-system/backend/internal/logagent"
	"ops-system/backend/internal/middleware"
	"ops-system/backend/internal/notify"
	"ops-system/backend/internal/repository"
	"ops-system/backend/internal/service"
	"ops-system/backend/internal/vm"

	"github.com/gin-gonic/gin"
	"github.com/google/uuid"
	"github.com/prometheus/client_golang/prometheus/promhttp"
	"go.uber.org/zap"
	"gorm.io/gorm"
)

// NewRouter 注册路由与全局中间件。
func NewRouter(cfg *config.Config, log *zap.Logger, db *gorm.DB) *gin.Engine {
	if cfg.Server.Mode == "release" {
		gin.SetMode(gin.ReleaseMode)
	} else if cfg.Server.Mode == "test" {
		gin.SetMode(gin.TestMode)
	}

	r := gin.New()
	r.Use(gin.Recovery())
	r.Use(middleware.RequestLogger(log))
	r.Use(middleware.CORS(cfg.CORS.AllowedOrigins))

	lim := middleware.NewIPRateLimiter(cfg.RateLimit.RequestsPerSecond, cfg.RateLimit.Burst)
	r.Use(lim.RateLimit())

	r.GET("/health", func(c *gin.Context) {
		c.JSON(http.StatusOK, gin.H{"status": "ok"})
	})
	r.GET("/metrics", gin.WrapH(promhttp.Handler()))

	api := r.Group("/api/v1")
	{
		api.GET("/health", func(c *gin.Context) {
			c.JSON(http.StatusOK, gin.H{"code": 0, "message": "success", "data": gin.H{"status": "ok", "version": "0.3.0"}})
		})
		if db != nil {
			api.GET("/health/db", func(c *gin.Context) {
				sqlDB, err := db.DB()
				if err != nil {
					c.JSON(http.StatusServiceUnavailable, gin.H{"code": http.StatusServiceUnavailable, "error": "INTERNAL_ERROR", "message": "database driver unavailable"})
					return
				}
				if err := sqlDB.PingContext(c.Request.Context()); err != nil {
					c.JSON(http.StatusServiceUnavailable, gin.H{"code": http.StatusServiceUnavailable, "error": "SERVICE_UNAVAILABLE", "message": "database ping failed"})
					return
				}
				c.JSON(http.StatusOK, gin.H{"code": 0, "message": "success", "data": gin.H{"status": "ok", "database": "postgresql"}})
			})

			workspaceRepo := repository.NewWorkspaceRepository(db)
			userRepo := repository.NewUserRepository(db)
			memberRepo := repository.NewWorkspaceMemberRepository(db)
			apiTokenRepo := repository.NewAPITokenRepository(db)
			vmRouteRepo := repository.NewVMRouteRepository(db)
			vmClusterRepo := repository.NewVMClusterRepository(db)
			datasourceRepo := repository.NewDatasourceRepository(db)
			provisioningTaskRepo := repository.NewProvisioningTaskRepository(db)
			auditLogRepo := repository.NewAuditLogRepository(db)
			instanceRepo := repository.NewInstanceRepository(db)
			alertRepo := repository.NewAlertRuleRepository(db)
			alertEventRepo := repository.NewAlertEventRepository(db)
			channelRepo := repository.NewNotificationChannelRepository(db)
			logInstanceRepo := repository.NewLogInstanceRepository(db)
			logClusterRepo := repository.NewLogClusterRepository(db)
			integrationTemplateRepo := repository.NewIntegrationTemplateRepository(db)
			integrationInstallRepo := repository.NewIntegrationInstallationRepository(db)
			metricRepo := repository.NewMetricRepository(db)
			grafanaInstanceRepo := repository.NewGrafanaInstanceRepository(db)
			grafanaInstanceSvc := service.NewGrafanaInstanceService(grafanaInstanceRepo)
			clusterRepo := repository.NewClusterRepository(db)
			zoneRepo := repository.NewZoneRepository(db)
			businessClusterRepo := repository.NewBusinessClusterRepository(db)

			auditSvc := service.NewAuditService(auditLogRepo)
			memberSvc := service.NewWorkspaceMemberService(memberRepo, workspaceRepo)
			userSvc := service.NewUserService(userRepo, memberRepo)
			apiTokenSvc := service.NewAPITokenService(apiTokenRepo, userRepo)
			authSvc := service.NewAuthService(userRepo, auditSvc, cfg.JWT.Secret, cfg.JWT.ExpireHours)
			authH := handler.NewAuthHandler(authSvc, userSvc, memberSvc, cfg.JWT.Secret)
			userH := handler.NewUserHandler(userSvc, memberSvc, cfg.JWT.Secret)
			memberH := handler.NewWorkspaceMemberHandler(memberSvc, userSvc)
			apiTokenH := handler.NewAPITokenHandler(apiTokenSvc)
			auditH := handler.NewAuditHandler(auditSvc)


			vmSync := vm.NewSyncService(&cfg.VM, log)
			vmQuery := vm.NewQueryClient(&cfg.VM)
			vmRoutes := vm.NewRouteBuilder(&cfg.VM)
			grafanaClient := grafana.NewClient(&cfg.Grafana, log)
			grafanaSvc := service.NewGrafanaService(grafanaClient, grafanaInstanceRepo, workspaceRepo, log)

			// Grafana resolver：按 Grafana 实例 id 动态构造 client；供 WorkspaceService 和 integration applier 共用。
			grafanaResolver := func(ctx context.Context, instanceID *uuid.UUID) (*grafana.Client, error) {
				if instanceID == nil {
					return grafanaClient, nil
				}
				grafanaInstance, err := grafanaInstanceRepo.GetByID(ctx, *instanceID)
				if err != nil {
					return grafanaClient, err
				}
				if grafanaInstance == nil || grafanaInstance.Status != "active" || strings.TrimSpace(grafanaInstance.URL) == "" {
					return grafanaClient, nil
				}
				subCfg := config.GrafanaConfig{
					Enabled:                 true,
					BaseURL:                 grafanaInstance.URL,
					APIKey:                  grafanaInstance.AdminTokenEnc,
					AdminUser:               grafanaInstance.AdminUser,
					AdminPassword:           grafanaInstance.AdminPassword,
					HTTPTimeoutSeconds:      cfg.Grafana.HTTPTimeoutSeconds,
					PrometheusDatasourceURL: cfg.Grafana.PrometheusDatasourceURL,
					OrgNamePrefix:           cfg.Grafana.OrgNamePrefix,
				}
				return grafana.NewClient(&subCfg, log), nil
			}

			orch, err := service.NewOrchestratorService(cfg, log)
			if err != nil {
				log.Fatal("orchestrator_init", zap.Error(err))
			}
			var (
				helmClient *helm.Client
				k8sClient  *k8s.Client
			)
			if cfg.Kubernetes.InCluster || strings.TrimSpace(cfg.Kubernetes.Kubeconfig) != "" {
				hc, herr := helm.NewClient(cfg.Kubernetes.Kubeconfig)
				if herr != nil {
					log.Warn("scale_helm_client_init_failed", zap.Error(herr))
				} else {
					helmClient = hc
				}
				kc, kerr := k8s.NewClient(cfg.Kubernetes.Kubeconfig, cfg.Kubernetes.InCluster)
				if kerr != nil {
					log.Warn("scale_k8s_client_init_failed", zap.Error(kerr))
				} else {
					k8sClient = kc
				}
			}
			vmOperator := vm.NewVMOperatorClient(k8sClient, &cfg.VM)
			tenantProvisioner := service.NewWorkspaceProvisioner(
				provisioningTaskRepo,
				vmRouteRepo,
				datasourceRepo,
				vmClusterRepo,
				auditLogRepo,
				vmRoutes,
				vmOperator,
				&cfg.VM,
				log,
			)
			workspaceSvc := service.NewWorkspaceService(workspaceRepo, instanceRepo, vmSync, vmQuery, tenantProvisioner, grafanaResolver, grafanaSvc, orch, log)
			workspaceH := handler.NewTenantHandler(workspaceSvc, userSvc)

			instanceSvc := service.NewInstanceService(instanceRepo, workspaceRepo, zoneRepo, vmClusterRepo, businessClusterRepo, integrationInstallRepo, orch, vmOperator, vmRoutes, vmQuery, log)
			instanceH := handler.NewInstanceHandler(instanceSvc, userSvc, grafanaInstanceSvc)
			k8sOps := service.NewK8sResourceOperator(k8sClient, log)
			platformScaleSvc := service.NewPlatformScaleService(k8sOps, vmClusterRepo)
			platformBootstrapSvc := service.NewPlatformBootstrapService(helmClient, k8sClient)
			var idemStore idempotency.Store
			redisAddr := fmt.Sprintf("%s:%d", cfg.Redis.Host, cfg.Redis.Port)
			if cfg.Redis.Host != "" && cfg.Redis.Port > 0 {
				redisStore := idempotency.NewRedisStore(redisAddr)
				if err := redisStore.Ping(context.Background()); err != nil {
					log.Warn("platform_idempotency_redis_unavailable", zap.String("addr", redisAddr), zap.Error(err))
				} else {
					idemStore = redisStore
				}
			}
			platformH := handler.NewPlatformHandler(platformScaleSvc, platformBootstrapSvc, log, idemStore, auditSvc)

			// grafanaSvc moved up(grafanaClient, grafanaInstanceRepo, workspaceRepo, log)
			grafanaH := handler.NewGrafanaHandler(grafanaSvc, log)
			notifySvc := notify.NewNotifyService(log)
			if cfg.Notification.Email.Enabled {
				notifySvc.Register(notify.NewEmailSender(
					cfg.Notification.Email.SMTPHost,
					cfg.Notification.Email.SMTPPort,
					cfg.Notification.Email.Username,
					cfg.Notification.Email.Password,
					cfg.Notification.Email.From,
				))
			}
			notifySvc.Register(notify.NewWebhookSender())
			notifySvc.Register(notify.NewSlackSender())
			notifySvc.Register(notify.NewSMSSender(log))

			alertSvc := service.NewAlertService(alertRepo, workspaceRepo, vmOperator, log)
			alertEventSvc := service.NewAlertEventService(alertEventRepo, alertRepo, channelRepo, notifySvc, log)
			webhookH := handler.NewWebhookHandler(alertEventSvc, cfg.Server.WebhookToken)
			channelSvc := service.NewNotificationChannelService(channelRepo, workspaceRepo, log)
			alertH := handler.NewAlertHandler(alertSvc, alertEventSvc, channelSvc, userSvc)

			logInstanceSvc := service.NewLogInstanceService(logInstanceRepo, logClusterRepo, zoneRepo, repository.NewLogSetRepository(db), &cfg.Logs)
			logInstanceH := handler.NewLogInstanceHandler(logInstanceSvc, userSvc)

			renderer := integrationpkg.NewRenderer()

			// 集群 resolver：按 cluster id 动态构造 k8s.Client；nil / 查询失败 / 创建失败 → 回退默认 client。
			// 用线程安全缓存避免每次 Apply 反复做 discovery（并发请求同 cluster 也安全）。
			clusterClientCache := k8s.NewClusterClientCache()
			k8sResolver := integrationpkg.K8sClientResolver(func(ctx context.Context, clusterID *uuid.UUID) (*k8s.Client, error) {
				if clusterID == nil {
					return k8sClient, nil
				}
				cluster, err := clusterRepo.GetByID(ctx, *clusterID)
				if err != nil {
					return k8sClient, err
				}
				if cluster == nil || cluster.Status != "active" {
					return k8sClient, nil
				}
				fp := fmt.Sprintf("%v|%s|%s", cluster.InCluster, cluster.KubeconfigPath, cluster.Kubeconfig)
				if cached := clusterClientCache.Get(*clusterID, fp); cached != nil {
					return cached, nil
				}
				kubeconfigPath := strings.TrimSpace(cluster.KubeconfigPath)
				inCluster := cluster.InCluster
				// 若只提供了 kubeconfig 原文，可在临时目录里写一个文件；当前实现仅支持 path/incluster，
				// inline kubeconfig 暂不自动落盘（可在后续加 helper）。
				if strings.TrimSpace(cluster.Kubeconfig) != "" && kubeconfigPath == "" {
					log.Warn("cluster_inline_kubeconfig_ignored",
						zap.String("cluster", cluster.Name),
						zap.String("hint", "supply kubeconfig_path instead"))
				}
				if !inCluster && kubeconfigPath == "" {
					return k8sClient, nil
				}
				cli, kerr := k8s.NewClient(kubeconfigPath, inCluster)
				if kerr != nil {
					log.Warn("cluster_k8s_client_init_failed",
						zap.String("cluster", cluster.Name),
						zap.Error(kerr))
					return k8sClient, kerr
				}
				clusterClientCache.Put(*clusterID, fp, cli)
				return cli, nil
			})

			var integrationApplier integrationpkg.Applier
			if k8sClient != nil || (grafanaClient != nil && grafanaClient.Enabled()) {
				integrationApplier = integrationpkg.NewCompositeApplier(
					k8sClient, k8sResolver,
					grafanaClient, integrationpkg.GrafanaClientResolver(grafanaResolver),
					log,
				)
				log.Info("integration_applier_enabled",
					zap.Bool("k8s", k8sClient != nil),
					zap.Bool("grafana", grafanaClient != nil && grafanaClient.Enabled()))
			} else {
				integrationApplier = integrationpkg.NewNoopApplier()
				log.Warn("integration_applier_noop", zap.String("reason", "k8s & grafana unavailable"))
			}
			_ = clusterRepo // 下方 cluster handler 也会用到
			integrationTemplateSvc := service.NewIntegrationTemplateService(integrationTemplateRepo, integrationInstallRepo)
			integrationInstallSvc := service.NewIntegrationInstallationService(
				integrationInstallRepo, integrationTemplateRepo, instanceRepo, renderer, integrationApplier,
			)
			integrationH := handler.NewIntegrationHandler(integrationTemplateSvc, integrationInstallSvc, userSvc)

			metricSvc := service.NewMetricService(metricRepo, integrationTemplateRepo)
			metricH := handler.NewMetricHandler(metricSvc)

			seeder := service.NewIntegrationSeeder(db, integrationTemplateRepo, metricRepo, log)
			if err := seeder.SeedBuiltin(context.Background()); err != nil {
				log.Warn("integration_seed_failed", zap.Error(err))
			}

			grafanaInstanceH := handler.NewGrafanaInstanceHandler(grafanaInstanceSvc)

			clusterSvc := service.NewClusterService(clusterRepo)
			clusterH := handler.NewClusterHandler(clusterSvc)

			zoneSvc := service.NewZoneService(zoneRepo, instanceRepo, vmClusterRepo, logClusterRepo, grafanaInstanceRepo, clusterRepo, clusterClientCache, &cfg.VM, &cfg.Logs, &cfg.Helm.Charts, log)
			zonePreflightSvc := service.NewZonePreflightService(zoneRepo, clusterRepo, clusterClientCache, &cfg.Helm, log)
			zoneComponentSvc := service.NewZoneComponentService(zoneRepo, vmClusterRepo, logClusterRepo, grafanaInstanceRepo)
			zoneH := handler.NewZoneHandler(zoneSvc, zonePreflightSvc, zoneComponentSvc)

			vectorAgent := logagent.NewVectorAgentClient()
			businessClusterSvc := service.NewBusinessClusterService(businessClusterRepo, instanceRepo, logInstanceRepo, workspaceRepo, zoneRepo, vmClusterRepo, logClusterRepo, vmOperator, vectorAgent, log)
			businessClusterH := handler.NewBusinessClusterHandler(businessClusterSvc, userSvc)

			entityRepo := repository.NewEntityRepository(db)
			metricSetRepo := repository.NewMetricSetRepository(db)
			logSetRepo := repository.NewLogSetRepository(db)
			dataLinkRepo := repository.NewDataLinkRepository(db)
			umodelSvc := service.NewUModelService(entityRepo, metricSetRepo, logSetRepo, dataLinkRepo)
			umodelH := handler.NewUModelHandler(umodelSvc, userSvc)

			api.POST("/auth/login", authH.Login)
			api.POST("/users/bootstrap", userH.Bootstrap)
			api.POST("/webhooks/alertmanager", webhookH.Alertmanager)

			// Grafana 反向代理（通过 cookie 识别目标实例，auth proxy 免手动登录）
			api.Any("/grafana/proxy/*path", grafanaInstanceH.ProxyGrafana)

			protected := api.Group("")
			protected.Use(middleware.JWTAuth(cfg.JWT.Secret, apiTokenSvc))
			protected.GET("/auth/me", authH.Me)


			tg := protected.Group("/workspaces")
			tg.GET("", workspaceH.List)
			tg.GET("/:id/metrics", workspaceH.Metrics)
			tg.GET("/:id", workspaceH.Get)
			tg.GET("/:id/members", memberH.List)
			tg.POST("/:id/members", memberH.Add)
			tg.PUT("/:id/members/:userId", memberH.Update)
			tg.DELETE("/:id/members/:userId", memberH.Remove)

			atg := protected.Group("/api-tokens")
			atg.GET("", apiTokenH.List)
			atg.POST("", apiTokenH.Create)
			atg.DELETE("/:id", apiTokenH.Revoke)

			ug := protected.Group("/users")
			ug.GET("", userH.List)
			ug.GET("/:id", userH.Get)
			ug.PUT("/:id", userH.Update)

		ig := protected.Group("/instances")
		ig.GET("", instanceH.List)
		ig.GET("/:id", instanceH.Get)
		ig.GET("/:id/metrics", instanceH.Metrics)
		ig.POST("/:id/login", instanceH.Login)

			gig := protected.Group("/grafana/instances/:instanceId")
			gig.GET("/orgs", grafanaH.ListOrgs)
			gig.GET("/orgs/:orgId/users", grafanaH.ListOrgUsers)
			gig.GET("/orgs/:orgId/datasources", grafanaH.ListDatasources)
			gig.GET("/orgs/:orgId/dashboards", grafanaH.ListDashboards)
			gig.GET("/orgs/:orgId/dashboards/:uid", grafanaH.GetDashboard)

			gig.GET("/plugins", grafanaH.ListPlugins)
			gig.GET("/health", grafanaH.HealthCheck)
			gig.GET("/admin/stats", grafanaH.AdminStats)
			gig.GET("/admin/settings", grafanaH.AdminSettings)

			ag := protected.Group("/alerts")
			ag.GET("/rules", alertH.ListRules)
			ag.GET("/events", alertH.ListEvents)
			ag.GET("/events/:id", alertH.GetEvent)
			ag.PUT("/events/:id/ack", alertH.AckEvent)
			ag.GET("/channels", alertH.ListChannels)
			ag.GET("/stats/summary", alertH.Summary)
			ag.GET("/stats/trend", alertH.Trend)
			ag.GET("/stats/by-level", alertH.StatsByLevel)
			ag.GET("/stats/by-rule", alertH.StatsByRule)

			// 日志实例（查询）
			lg := protected.Group("/log-instances")
			lg.GET("", logInstanceH.List)
			lg.GET("/:id", logInstanceH.Get)
			lg.POST("/:id/query", logInstanceH.Query)

			// 接入中心（查询）
			intg := protected.Group("/integrations")
			intg.GET("/categories", integrationH.ListCategories)
			intg.GET("/templates", integrationH.ListTemplates)
			intg.GET("/templates/:id", integrationH.GetTemplate)
			intg.GET("/templates/:id/versions", integrationH.ListVersions)
			intg.POST("/install/plan", integrationH.InstallPlan)
			intg.POST("/install", integrationH.Install)
			intg.GET("/installations", integrationH.ListInstallations)
			intg.GET("/installations/:id", integrationH.GetInstallation)
			intg.GET("/installations/:id/revisions", integrationH.ListInstallationRevisions)
			intg.POST("/installations/:id/upgrade", integrationH.Upgrade)
			intg.POST("/installations/:id/rollback", integrationH.Rollback)
			intg.DELETE("/installations/:id", integrationH.Uninstall)

			// 指标库（查询）
			mg := protected.Group("/metrics")
			mg.GET("", metricH.List)
			mg.GET("/:id", metricH.Get)
			mg.GET("/:id/related", metricH.Related)

			// Grafana 纳管实例（查询）
			ghg := protected.Group("/grafana/instances")
			ghg.GET("", grafanaInstanceH.List)
			ghg.GET("/:instanceId", grafanaInstanceH.Get)
			ghg.POST("/:instanceId/login", grafanaInstanceH.Login)

			// K8s 集群（查询）
			cg := protected.Group("/clusters")
			cg.GET("", clusterH.List)
			cg.GET("/:id", clusterH.Get)

			zg := protected.Group("/zones")
			zg.GET("", zoneH.List)
			zg.GET("/:id", zoneH.Get)
			zg.GET("/:id/preflight", zoneH.Preflight)
			zg.GET("/:id/components", zoneH.Components)

			bg := protected.Group("/business-clusters")
			bg.GET("", businessClusterH.List)
			bg.GET("/:id", businessClusterH.Get)

			umg := protected.Group("/umodel")
			umg.GET("/entities", umodelH.ListEntities)
			umg.POST("/entities", umodelH.CreateEntity)
			umg.DELETE("/entities/:id", umodelH.DeleteEntity)
			umg.GET("/entities/:id/data-links", umodelH.ListDataLinks)
			umg.GET("/metric-sets", umodelH.ListMetricSets)
			umg.POST("/metric-sets", umodelH.CreateMetricSet)
			umg.DELETE("/metric-sets/:id", umodelH.DeleteMetricSet)
			umg.GET("/log-sets", umodelH.ListLogSets)
			umg.POST("/log-sets", umodelH.CreateLogSet)
			umg.DELETE("/log-sets/:id", umodelH.DeleteLogSet)
			umg.POST("/data-links", umodelH.CreateDataLink)
			umg.DELETE("/data-links/:id", umodelH.DeleteDataLink)

			admin := protected.Group("")
			admin.Use(middleware.RequireRole("admin"))

			admin.GET("/audits", auditH.List)


			adminTG := admin.Group("/workspaces")
			adminTG.POST("", workspaceH.Create)
			adminTG.PUT("/:id", workspaceH.Update)
			adminTG.DELETE("/:id", workspaceH.Delete)

			adminUG := admin.Group("/users")
			adminUG.POST("", userH.Create)
			adminUG.DELETE("/:id", userH.Delete)

			adminIG := admin.Group("/instances")
			adminIG.POST("", instanceH.Create)
			adminIG.PUT("/:id", instanceH.Update)
			adminIG.DELETE("/:id", instanceH.Delete)

			adminGI := admin.Group("/grafana/instances/:instanceId")
			adminGI.POST("/orgs", grafanaH.CreateOrg)
			adminGI.PUT("/orgs/:orgId", grafanaH.UpdateOrg)
			adminGI.DELETE("/orgs/:orgId", grafanaH.DeleteOrg)
			adminGI.POST("/orgs/:orgId/users", grafanaH.AddOrgUser)
			adminGI.DELETE("/orgs/:orgId/users/:userId", grafanaH.RemoveOrgUser)
			adminGI.POST("/orgs/:orgId/datasources", grafanaH.CreateDatasource)
			adminGI.DELETE("/orgs/:orgId/datasources/:dsId", grafanaH.DeleteDatasource)
			adminGI.PUT("/orgs/:orgId/datasources/:dsId", grafanaH.UpdateDatasource)
			adminGI.POST("/orgs/:orgId/datasources/test", grafanaH.TestDatasource)
			adminGI.POST("/orgs/:orgId/dashboards/import", grafanaH.ImportDashboard)
			adminGI.DELETE("/orgs/:orgId/dashboards/:uid", grafanaH.DeleteDashboard)

			adminGI.POST("/plugins/:pluginId/install", grafanaH.InstallPlugin)
			adminGI.DELETE("/plugins/:pluginId", grafanaH.UninstallPlugin)

			adminAG := admin.Group("/alerts")
			adminAG.POST("/rules", alertH.CreateRule)
			adminAG.PUT("/rules/:id", alertH.UpdateRule)
			adminAG.DELETE("/rules/:id", alertH.DeleteRule)
			adminAG.POST("/channels", alertH.CreateChannel)
			adminAG.PUT("/channels/:id", alertH.UpdateChannel)
			adminAG.DELETE("/channels/:id", alertH.DeleteChannel)

			adminPG := admin.Group("/platform/scaling")
			adminPG.POST("/bootstrap/shared/init", platformH.InitSharedCluster)
			adminPG.GET("/audits", platformH.ListAudits)
			adminPG.GET("/vmcluster/targets", platformH.ListVMClusterTargets)
			adminPG.POST("/vmcluster", platformH.ScaleVMCluster)

			// 日志实例（写）
			adminLG := admin.Group("/log-instances")
			adminLG.POST("", logInstanceH.Create)
			adminLG.PUT("/:id", logInstanceH.Update)
			adminLG.DELETE("/:id", logInstanceH.Delete)

			// 接入中心模版（写）
			adminINTG := admin.Group("/integrations")
			adminINTG.POST("/templates", integrationH.CreateTemplate)
			adminINTG.PUT("/templates/:id", integrationH.UpdateTemplate)
			adminINTG.DELETE("/templates/:id", integrationH.DeleteTemplate)
			adminINTG.POST("/templates/:id/versions", integrationH.CreateVersion)
			adminINTG.DELETE("/templates/:id/versions/:version", integrationH.DeleteVersion)

			// 指标库（写）
			adminMG := admin.Group("/metrics")
			adminMG.POST("", metricH.Create)
			adminMG.PUT("/:id", metricH.Update)
			adminMG.DELETE("/:id", metricH.Delete)
			adminMG.POST("/reparse/:templateId", metricH.Reparse)

			// Grafana 纳管实例（写）
			adminGHG := admin.Group("/grafana/instances")
			adminGHG.POST("", grafanaInstanceH.Create)
			adminGHG.PUT("/:instanceId", grafanaInstanceH.Update)
			adminGHG.DELETE("/:instanceId", grafanaInstanceH.Delete)

			// K8s 集群（写）
			adminCG := admin.Group("/clusters")
			adminCG.POST("", clusterH.Create)
			adminCG.PUT("/:id", clusterH.Update)
			adminCG.DELETE("/:id", clusterH.Delete)

			adminZG := admin.Group("/zones")
			adminZG.POST("", zoneH.Create)
			adminZG.PUT("/:id", zoneH.Update)
			adminZG.DELETE("/:id", zoneH.Delete)
			adminZG.POST("/:id/init-shared", zoneH.InitShared)
			adminZG.POST("/:id/init-logs", zoneH.InitLogs)
			adminZG.POST("/:id/init-grafana", zoneH.InitGrafana)

			adminBG := admin.Group("/business-clusters")
			adminBG.POST("", businessClusterH.Create)
			adminBG.POST("/:id/enable-logs", businessClusterH.EnableLogs)
			adminBG.POST("/:id/disable-logs", businessClusterH.DisableLogs)
			adminBG.DELETE("/:id", businessClusterH.Delete)
		}
	}

	return r
}