# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: ops-system API
  version: 0.5.0-logs
  description: SaaS 可观测控制面 — 监控 + 日志数据面（Vector→Kafka→Aggregator→VL）
paths:
...
  /api/v1/log-instances:
    get:
      summary: 列出日志工作空间
      tags: [log-instances]
    post:
      summary: 创建日志工作空间
      tags: [log-instances]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [tenant_id, zone_id, instance_name]
              properties:
                tenant_id: { type: string, format: uuid }
                zone_id: { type: string, format: uuid }
                instance_name: { type: string }
                backend_type: { enum: [victorialogs], default: victorialogs }
                retention_days: { type: integer }
  /api/v1/log-instances/{id}/query:
    post:
      summary: LogsQL 查询（强制租户过滤）
      tags: [log-instances]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                query: { type: string }
                start: { type: string, format: date-time }
                end: { type: string, format: date-time }
                limit: { type: integer, maximum: 1000 }
  /api/v1/instances:
    post:
      summary: 创建 Metric Workspace（shared-only）
      tags: [instances]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [workspace_id, zone_id, instance_name, instance_type]
              properties:
                workspace_id: { type: string, format: uuid }
                zone_id: { type: string, format: uuid }
                instance_name: { type: string }
                instance_type: { enum: [metrics] }
                template_type: { enum: [shared], default: shared }
...
  /api/v1/business-clusters/{id}/enable-logs:
    post:
      summary: 启用 Vector Agent 日志采集（→ Kafka）
      tags: [business-clusters]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [log_instance_id]
              properties:
                log_instance_id: { type: string, format: uuid }
...